Verify stats after seekAsync
