+++ pulsar-broker/src/test/java/org/apache/pulsar/client/api/MultiTopicsConsumerTest.java
@@ -29,22 +29,16 @@ import static org.testng.Assert.assertTrue;
 
 import com.google.common.collect.Lists;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.TreeSet;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import lombok.Cleanup;
@@ -376,67 +370,5 @@ public class MultiTopicsConsumerTest extends ProducerConsumerBase {
                 .subscriptionName("sub").subscribe();
         assertTrue(consumer instanceof MultiTopicsConsumerImpl);
         assertTrue(consumer.isConnected());
-    }
-
-    @DataProvider
-    public static Object[][] seekByFunction() {
-        return new Object[][] {
-                { true }, { false }
-        };
-    }
-
-    @Test(timeOut = 30000, dataProvider = "seekByFunction")
-    public void testSeekToNewerPosition(boolean seekByFunction) throws Exception {
-        final var topic1 = TopicName.get(newTopicName()).toString();
-        final var topic2 = TopicName.get(newTopicName()).toString();
-        @Cleanup final var producer1 = pulsarClient.newProducer(Schema.STRING).topic(topic1).create();
-        @Cleanup final var producer2 = pulsarClient.newProducer(Schema.STRING).topic(topic2).create();
-        producer1.send("1-0");
-        producer2.send("2-0");
-        producer1.send("1-1");
-        producer2.send("2-1");
-        final var consumer1 = pulsarClient.newConsumer(Schema.STRING)
-                .topics(Arrays.asList(topic1, topic2)).subscriptionName("sub")
-                .subscriptionInitialPosition(SubscriptionInitialPosition.Earliest).subscribe();
-        final var timestamps = new ArrayList<Long>();
-        for (int i = 0; i < 4; i++) {
-            timestamps.add(consumer1.receive().getPublishTime());
-        }
-        timestamps.sort(Comparator.naturalOrder());
-        final var timestamp = timestamps.get(2);
-        consumer1.close();
-
-        final Function<Consumer<String>, CompletableFuture<Void>> seekAsync = consumer ->
-                seekByFunction ? consumer.seekAsync(__ -> timestamp) : consumer.seekAsync(timestamp);
-
-        @Cleanup final var consumer2 = pulsarClient.newConsumer(Schema.STRING)
-                .topics(Arrays.asList(topic1, topic2)).subscriptionName("sub-2")
-                .subscriptionInitialPosition(SubscriptionInitialPosition.Earliest).subscribe();
-        seekAsync.apply(consumer2).get();
-        final var values = new TreeSet<String>();
-        for (int i = 0; i < 2; i++) {
-            values.add(consumer2.receive().getValue());
-        }
-        assertEquals(values, new TreeSet<>(Arrays.asList("1-1", "2-1")));
-
-        final var valuesInListener = new CopyOnWriteArrayList<String>();
-        @Cleanup final var consumer3 = pulsarClient.newConsumer(Schema.STRING)
-                .topics(Arrays.asList(topic1, topic2)).subscriptionName("sub-3")
-                .messageListener((MessageListener<String>) (__, msg) -> valuesInListener.add(msg.getValue()))
-                .subscriptionInitialPosition(SubscriptionInitialPosition.Earliest).subscribe();
-        seekAsync.apply(consumer3).get();
-        if (valuesInListener.isEmpty()) {
-            Awaitility.await().untilAsserted(() -> assertEquals(valuesInListener.size(), 2));
-            assertEquals(valuesInListener.stream().sorted().toList(), Arrays.asList("1-1", "2-1"));
-        } // else: consumer3 has passed messages to the listener before seek, in this case we cannot assume anything
-
-        @Cleanup final var consumer4 = pulsarClient.newConsumer(Schema.STRING)
-                .topics(Arrays.asList(topic1, topic2)).subscriptionName("sub-4")
-                .subscriptionInitialPosition(SubscriptionInitialPosition.Earliest).subscribe();
-        seekAsync.apply(consumer4).get();
-        final var valuesInReceiveAsync = new ArrayList<String>();
-        valuesInReceiveAsync.add(consumer4.receiveAsync().get().getValue());
-        valuesInReceiveAsync.add(consumer4.receiveAsync().get().getValue());
-        assertEquals(valuesInReceiveAsync.stream().sorted().toList(), Arrays.asList("1-1", "2-1"));
     }
 }
